requirements.txt
requests>=2.31.0
playwright>=1.44.0
Pillow>=10.0.0
apscheduler>=3.10.4
anthropic>=0.40.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0